# openapi/openapi.yml
openapi: 3.0.3
info:
  title: LMS Core API
  version: 0.2.10
  description: Core API for the Learning Management System. Management endpoints under `/admin/**` require the ADMIN role. Authenticated self-service endpoints live under `/private/**`.
  contact:
    name: Sang Dang
    email: contact@fundamentals.io.vn
...
security:
  - cognitoOAuth:
      - openid
      - email
      - profile
  - bearerAuth: []
paths:
...
  /admin/classrooms/{classroomId}/members/{memberId}/attendances/query:
    post:
      tags:
        - ClassroomSessions
      operationId: getClassroomMemberAttendances
      summary: Get classroom session attendances for a member
      description: >
        Query a classroom member's session attendance history. Use `attendanceDate` filters (for example `gte` / `lte`) to limit the range, and `status` (for example `ATTENDED`) to see sessions the member joined. Pagination and sorting follow the same pattern as other query endpoints.

      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomSessionAttendanceFilter"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllClassroomSessionAttendancesResponse"
        "404":
          description: Classroom or classroom member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/classrooms/{classroomId}/sessions:
    post:
      tags:
        - ClassroomSessions
      operationId: createClassroomSession
      summary: Create an adhoc classroom session
      description: >
        Creates an adhoc session for a classroom. Status is always OPEN and type is always ADHOC.
      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassroomSessionAdhocCommand"
      responses:
        "201":
          description: Classroom session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomSessionResponse"
        "400":
          description: Invalid create command
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Classroom not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/classrooms/{classroomId}/sessions/query:
    post:
      tags:
        - ClassroomSessions
      operationId: getAllClassroomSessions
      summary: Get all classroom sessions
      description: Query sessions belonging to a classroom.
      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassroomSessionFilter"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllClassroomSessionsResponse"
        "404":
          description: Classroom not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/classrooms/{classroomId}/sessions/{sessionId}:
    get:
      tags:
        - ClassroomSessions
      operationId: getClassroomSessionById
      summary: Get a classroom session by id
      description: Returns a single classroom session by its identifier.
      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Classroom session found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomSessionResponse"
        "404":
          description: Classroom session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/classrooms/{classroomId}/sessions/{sessionId}/complete:
    post:
      tags:
        - ClassroomSessions
      operationId: completeClassroomSession
      summary: Complete a classroom session
      description: >
        Moves the session status from OPEN to COMPLETED. COMPLETED and CANCELLED sessions are
        rejected.
      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Classroom session completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomSessionResponse"
        "400":
          description: Session is not OPEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Classroom session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/classrooms/{classroomId}/sessions/{sessionId}/cancel:
    post:
      tags:
        - ClassroomSessions
      operationId: cancelClassroomSession
      summary: Cancel a classroom session
      description: >
        Moves the session status from OPEN to CANCELLED. COMPLETED and CANCELLED sessions are
        rejected.
      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Classroom session cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomSessionResponse"
        "400":
          description: Session is not OPEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Classroom session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/classrooms/{classroomId}/sessions/{sessionId}/attendances:
    get:
      tags:
        - ClassroomSessions
      operationId: getAllClassroomSessionAttendances
      summary: Get all classroom session attendances
      description: Returns every attendance record for a classroom session. Not paginated.
      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllClassroomSessionAttendancesResponse"
        "404":
          description: Classroom session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - ClassroomSessions
      operationId: createClassroomSessionAttendances
      summary: Create classroom session attendances
      description: >
        Records attendance for up to 30 classroom members in a session. Duplicate classroom member
        ids in the request are rejected. If any member already has attendance for the session, the
        entire request fails. Session must be OPEN; COMPLETED and CANCELLED are rejected.

      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassroomSessionAttendancesCommand"
      responses:
        "201":
          description: Classroom session attendances created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllClassroomSessionAttendancesResponse"
        "400":
          description: Invalid create command, or session is not OPEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Classroom, session, or member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Attendance already exists for a member in this session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/classrooms/{classroomId}/sessions/{sessionId}/attendances/{attendanceId}:
    put:
      tags:
        - ClassroomSessions
      operationId: updateClassroomSessionAttendance
      summary: Update a classroom session attendance
      description: >
        Updates only the attendance status (present or absent) for an existing classroom session
        attendance record. Session must be OPEN; COMPLETED and CANCELLED are rejected.

      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/AttendanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassroomSessionAttendanceCommand"
      responses:
        "200":
          description: Classroom session attendance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomSessionAttendanceResponse"
        "400":
          description: Invalid update command, or session is not OPEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Classroom session attendance not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - ClassroomSessions
      operationId: deleteClassroomSessionAttendance
      summary: Delete a classroom session attendance
      description: >
        Hard-deletes a classroom session attendance record. Session must be OPEN; COMPLETED and
        CANCELLED are rejected.
      parameters:
        - $ref: "#/components/parameters/ClassroomIdPath"
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/AttendanceId"
      responses:
        "204":
          description: Classroom session attendance deleted
        "400":
          description: Session is not OPEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Classroom session attendance not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cognitoOAuth:
      type: oauth2
      description: >
        Cognito Hosted UI (authorization code + PKCE). Click Authorize in Swagger UI to open the Cognito login page. OAuth URLs are injected at runtime from COGNITO_DOMAIN (or COGNITO_AUTHORIZATION_URI / COGNITO_TOKEN_URI) — see application.yml under app.cognito. Client id comes from COGNITO_APP_CLIENT_ID. Cognito must allowlist http://localhost:8080/swagger-ui/oauth2-redirect.html (or your API origin) as a callback URL.

      flows:
        authorizationCode:
          # Replaced at serve time by OpenApiSpecController from app.cognito settings.
          authorizationUrl: __COGNITO_DOMAIN__/oauth2/authorize
          tokenUrl: __COGNITO_DOMAIN__/oauth2/token
          scopes:
            openid: OpenID Connect
            email: Email address
            profile: User profile
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Paste a Cognito JWT access token manually (alternative to Hosted UI login). Admin endpoints require the ADMIN Cognito group.
  parameters:
...
    ClassroomIdPath:
      name: classroomId
      in: path
      required: true
      description: Classroom identifier
      schema:
        type: string
        format: uuid
        description: Classroom UUID
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    MemberId:
      name: memberId
      in: path
      required: true
      description: Classroom member identifier
      schema:
        type: string
        format: uuid
        description: Classroom member UUID
        example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Classroom session identifier
      schema:
        type: string
        format: uuid
        description: Classroom session UUID
        example: 9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d
...
    AttendanceId:
      name: attendanceId
      in: path
      required: true
      description: Classroom session attendance identifier
      schema:
        type: string
        format: uuid
        description: Classroom session attendance UUID
        example: 1b9d6bcd-bbfd-4b2d-9b5d-ab8dfbbd4bed
...
  schemas:
...
    ClassroomSessionAttendanceStatus:
      type: string
      description: Attendance status for a classroom session
      enum:
        - ATTENDED
        - ABSENT
        - UNSET
      example: UNSET
    ClassroomSessionStatus:
      type: string
      description: Lifecycle status of a classroom session
      enum:
        - OPEN
        - COMPLETED
        - CANCELLED
      example: OPEN
    ClassroomSessionType:
      type: string
      description: How the classroom session was created
      enum:
        - ADHOC
        - SCHEDULE
      example: SCHEDULE
    CreateClassroomSessionAdhocCommand:
      type: object
      description: >
        Command to create an adhoc classroom session. Status is always OPEN and type is always ADHOC.
      additionalProperties: false
      required:
        - sessionDate
        - startTime
        - endTime
      properties:
        sessionDate:
          type: string
          format: date
          description: Calendar date of the classroom session
          example: "2026-07-19"
        startTime:
          type: string
          format: time
          description: Session start time
          example: "09:00:00"
        endTime:
          type: string
          format: time
          description: Session end time
          example: "10:30:00"
        name:
          type: string
          description: Optional display name of the classroom session
          minLength: 1
          maxLength: 255
          example: Week 1 lecture
        description:
          type: string
          description: Optional description of the classroom session
          minLength: 1
          maxLength: 2000
          example: Introduction to the course and classroom expectations
...
    ClassroomSessionFilter:
      type: object
      description: Filter for querying classroom sessions
      additionalProperties: false
      properties:
        page:
          type: integer
          default: 0
          minimum: 0
        size:
          type: integer
          default: 10
          minimum: 1
          maximum: 50
        sortBy:
          type: string
          pattern: '^(sessionDate|startTime|endTime|name|createdDate|status|type)$'
        sortDirection:
          type: string
          enum: [ASC, DESC]
          default: ASC
        filters:
          type: array
          items:
            type: object
            required: [field, operator, value]
            properties:
              field:
                type: string
                pattern: '^(sessionDate|startTime|endTime|name|status|type)$'
              operator:
                type: string
                pattern: '^(eq|like|gt|lt|gte|lte)$'
              value:
                type: string
    ClassroomSessionResponse:
      type: object
      description: Classroom session resource returned by the API
      additionalProperties: false
      required:
        - id
        - classroomId
        - sessionDate
        - startTime
        - endTime
        - status
        - type
        - createdDate
        - lastModifiedDate
      properties:
        id:
          type: string
          format: uuid
          description: Unique classroom session identifier
          example: 9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d
        classroomId:
          type: string
          format: uuid
          description: Classroom this session belongs to
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        sessionDate:
          type: string
          format: date
          description: Calendar date of the classroom session
          example: "2026-07-19"
        startTime:
          type: string
          format: time
          description: Session start time
          example: "09:00:00"
        endTime:
          type: string
          format: time
          description: Session end time
          example: "10:30:00"
        name:
          type: string
          description: Optional display name of the classroom session
          minLength: 1
          maxLength: 255
          example: Week 1 lecture
        description:
          type: string
          description: Optional description of the classroom session
          minLength: 1
          maxLength: 2000
          example: Introduction to the course and classroom expectations
        status:
          $ref: "#/components/schemas/ClassroomSessionStatus"
        type:
          $ref: "#/components/schemas/ClassroomSessionType"
        createdDate:
          type: string
          format: date-time
          description: When the session was created
          example: 2026-07-14T00:00:00Z
        lastModifiedDate:
          type: string
          format: date-time
          description: When the session was last modified
          example: 2026-07-14T00:00:00Z
    GetAllClassroomSessionsResponse:
      type: array
      description: Collection of classroom sessions
      items:
        $ref: "#/components/schemas/ClassroomSessionResponse"
    CreateClassroomSessionAttendanceCommand:
      type: object
      description: Command to create a classroom session attendance record
      additionalProperties: false
      required:
        - classroomMemberId
        - status
      properties:
        classroomMemberId:
          type: string
          format: uuid
          description: Classroom member whose attendance is being recorded
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        attendanceDate:
          type: string
          format: date-time
          description: When attendance was recorded; defaults to now when omitted
          example: 2026-07-19T09:05:00Z
        status:
          $ref: "#/components/schemas/ClassroomSessionAttendanceStatus"
    CreateClassroomSessionAttendancesCommand:
      type: object
      description: Command to create up to 30 classroom session attendance records in one request
      additionalProperties: false
      required:
        - attendances
      properties:
        attendances:
          type: array
          minItems: 1
          maxItems: 30
          items:
            $ref: "#/components/schemas/CreateClassroomSessionAttendanceCommand"
    UpdateClassroomSessionAttendanceCommand:
      type: object
      description: Command to update a classroom session attendance status
      additionalProperties: false
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/ClassroomSessionAttendanceStatus"
    ClassroomSessionAttendanceFilter:
      type: object
      description: Filter for querying a classroom member's session attendance history
      additionalProperties: false
      properties:
        page:
          type: integer
          default: 0
          minimum: 0
        size:
          type: integer
          default: 10
          minimum: 1
          maximum: 50
        sortBy:
          type: string
          pattern: '^(attendanceDate|status|createdDate)$'
        sortDirection:
          type: string
          enum: [ASC, DESC]
          default: ASC
        filters:
          type: array
          items:
            type: object
            required: [field, operator, value]
            properties:
              field:
                type: string
                pattern: '^(attendanceDate|status|sessionId)$'
              operator:
                type: string
                pattern: '^(eq|like|gt|lt|gte|lte)$'
              value:
                type: string
    GetAllClassroomSessionAttendancesResponse:
      type: array
      description: Collection of classroom session attendance records
      items:
        $ref: "#/components/schemas/ClassroomSessionAttendanceResponse"
    ClassroomSessionAttendanceResponse:
      type: object
      description: Classroom session attendance resource returned by the API
      additionalProperties: false
      required:
        - id
        - sessionId
        - classroomMemberId
        - status
        - createdDate
        - lastModifiedDate
      properties:
        id:
          type: string
          format: uuid
          description: Unique attendance identifier
          example: 1b9d6bcd-bbfd-4b2d-9b5d-ab8dfbbd4bed
        sessionId:
          type: string
          format: uuid
          description: Session this attendance belongs to
          example: 9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d
        classroomMemberId:
          type: string
          format: uuid
          description: Classroom member whose attendance was recorded
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        attendanceDate:
          type: string
          format: date-time
          description: When attendance was recorded
          example: 2026-07-19T09:05:00Z
        status:
          $ref: "#/components/schemas/ClassroomSessionAttendanceStatus"
        createdDate:
          type: string
          format: date-time
          description: When the attendance record was created
          example: 2026-07-14T00:00:00Z
        lastModifiedDate:
          type: string
          format: date-time
          description: When the attendance record was last modified
          example: 2026-07-14T00:00:00Z
...
    ErrorResponse:
      type: object
      description: Standard API error payload
      additionalProperties: false
      required:
        - code
        - message
        - status
        - timestamp
        - path
      properties:
        code:
          type: string
          description: Machine-readable error code
          minLength: 1
          maxLength: 64
          example: CLASSROOM_NOT_FOUND
        message:
          type: string
          description: Human-readable error message
          minLength: 1
          maxLength: 1024
          example: Classroom not found
        status:
          type: integer
          description: HTTP status code
          example: 404
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: 2026-07-16T00:00:00Z
        path:
          type: string
          description: Request path that caused the error
          minLength: 1
          maxLength: 2048
          example: /admin/classrooms/3fa85f64-5717-4562-b3fc-2c963f66afa6