# src/main/java/tech/sangdang/lmscoreapi/modules/management/api/ClassroomSessionController.java
package tech.sangdang.lmscoreapi.modules.management.api;

import java.util.UUID;
import lombok.RequiredArgsConstructor;
import org.jspecify.annotations.NonNull;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import tech.sangdang.lmscoreapi.generated.api.ClassroomSessionsApi;
import tech.sangdang.lmscoreapi.generated.model.*;
import tech.sangdang.lmscoreapi.modules.management.app.ClassroomSessionService;

@RestController
@RequiredArgsConstructor
public class ClassroomSessionController implements ClassroomSessionsApi {

  private final ClassroomSessionService classroomSessionService;

  @Override
  public ResponseEntity<?> createClassroomSession(
      @NonNull UUID classroomId,
      @NonNull CreateClassroomSessionAdhocCommand createClassroomSessionCommand) {
    return ResponseEntity.status(HttpStatus.CREATED)
        .body(
            classroomSessionService.createClassroomSessionAdhoc(
                classroomId, createClassroomSessionCommand));
  }

  @Override
  public ResponseEntity<?> getAllClassroomSessions(
      @NonNull UUID classroomId, @NonNull ClassroomSessionFilter classroomSessionFilter) {
    return ResponseEntity.ok(
        classroomSessionService.queryClassroomSessions(classroomId, classroomSessionFilter));
  }

  @Override
  public ResponseEntity<?> getClassroomMemberAttendances(
      @NonNull UUID classroomId,
      @NonNull UUID memberId,
      @NonNull ClassroomSessionAttendanceFilter classroomSessionAttendanceFilter) {
    return ResponseEntity.ok(
        classroomSessionService.queryClassroomSessionAttendancesByMember(
            classroomId, memberId, classroomSessionAttendanceFilter));
  }

  @Override
  public ResponseEntity<?> getClassroomSessionById(
      @NonNull UUID classroomId, @NonNull UUID sessionId) {
    return ResponseEntity.ok(
        classroomSessionService.getClassroomSessionById(classroomId, sessionId));
  }

  @Override
  public ResponseEntity<?> completeClassroomSession(
      @NonNull UUID classroomId, @NonNull UUID sessionId) {
    return ResponseEntity.ok(
        classroomSessionService.completeClassroomSession(classroomId, sessionId));
  }

  @Override
  public ResponseEntity<?> cancelClassroomSession(
      @NonNull UUID classroomId, @NonNull UUID sessionId) {
    return ResponseEntity.ok(
        classroomSessionService.cancelClassroomSession(classroomId, sessionId));
  }

  @Override
  public ResponseEntity<?> createClassroomSessionAttendances(
      @NonNull UUID classroomId,
      @NonNull UUID sessionId,
      @NonNull CreateClassroomSessionAttendancesCommand createClassroomSessionAttendancesCommand) {
    return ResponseEntity.status(HttpStatus.CREATED)
        .body(
            classroomSessionService.createClassroomSessionAttendances(
                classroomId, sessionId, createClassroomSessionAttendancesCommand));
  }

  @Override
  public ResponseEntity<?> getAllClassroomSessionAttendances(
      @NonNull UUID classroomId, @NonNull UUID sessionId) {
    return ResponseEntity.ok(
        classroomSessionService.getClassroomSessionAttendances(classroomId, sessionId));
  }

  @Override
  public ResponseEntity<?> updateClassroomSessionAttendance(
      @NonNull UUID classroomId,
      @NonNull UUID sessionId,
      @NonNull UUID attendanceId,
      @NonNull UpdateClassroomSessionAttendanceCommand updateClassroomSessionAttendanceCommand) {
    return ResponseEntity.ok(
        classroomSessionService.updateClassroomSessionAttendance(
            classroomId, sessionId, attendanceId, updateClassroomSessionAttendanceCommand));
  }

  @Override
  public ResponseEntity<?> deleteClassroomSessionAttendance(
      @NonNull UUID classroomId, @NonNull UUID sessionId, @NonNull UUID attendanceId) {
    classroomSessionService.deleteClassroomSessionAttendance(classroomId, sessionId, attendanceId);
    return ResponseEntity.noContent().build();
  }
}